Move mockConsole into util/test/ like mockCanvas
