+++ packages/ag-charts-community/src/chart/agChartV2.test.ts
@@ -9,7 +9,6 @@ import type {
 import { AgCharts } from './agChartV2';
 import type { Chart } from './chart';
 import * as examples from './test/examples';
-import { setupMockConsole } from './test/mockConsole';
 import type { TestCase } from './test/utils';
 import {
     IMAGE_SNAPSHOT_DEFAULTS,
@@ -18,6 +17,7 @@ import {
     prepareTestOptions,
     repeat,
     setupMockCanvas,
+    setupMockConsole,
     waitForChartStability,
 } from './test/utils';
