+++ packages/ag-charts-community/src/chart/chartHeapMemory.test.ts
@@ -5,8 +5,7 @@ import { memoryUsage } from 'process';
 import { AgCharts } from './agChartV2';
 import type { Chart } from './chart';
 import type { AgChartProxy } from './chartProxy';
-import { setupMockConsole } from './test/mockConsole';
-import { deproxy, prepareTestOptions, setupMockCanvas, waitForChartStability } from './test/utils';
+import { deproxy, prepareTestOptions, setupMockCanvas, setupMockConsole, waitForChartStability } from './test/utils';
 
 expect.extend({ toMatchImageSnapshot });
