+++ packages/ag-charts-community/src/chart/series/seriesLabels.test.ts
@@ -4,7 +4,6 @@ import { toMatchImageSnapshot } from 'jest-image-snapshot';
 import type { AgChartOptions } from '../../options/agChartOptions';
 import { AgCharts } from '../agChartV2';
 import type { Chart } from '../chart';
-import { setupMockConsole } from '../test/mockConsole';
 import type { TestCase } from '../test/utils';
 import {
     IMAGE_SNAPSHOT_DEFAULTS,
@@ -14,6 +13,7 @@ import {
     prepareTestOptions,
     repeat,
     setupMockCanvas,
+    setupMockConsole,
     waitForChartStability,
 } from '../test/utils';
 import * as examples from './test/examples';
