+++ packages/ag-charts-community/src/main-test.ts
@@ -1,6 +1,5 @@
 // Undocumented APIs used by Enterprise Modules.
 export * from './chart/test/utils';
-export * from './chart/test/mockConsole';
 export { COMMUNITY_AND_ENTERPRISE_EXAMPLES as GALLERY_EXAMPLES } from './chart/test/examples-gallery';
 export * from './chart/test/examples';
 export * from './chart/series/test/examples';
